requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0 
pycryptodome==3.19.0
pythonnet==3.0.5
python-dotenv==1.0.0
UnityPy==1.20.17 